requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1

# Strands Agents SDK
strands-agents>=1.0.0

# Model Providers
ollama>=0.1.0
openai>=1.0.0

# HTTP Client
httpx==0.28.1
requests==2.32.3

# JWT Authentication
PyJWT[crypto]==2.10.1
cryptography==44.0.0

# Spell Checking (Spanish)
pyspellchecker==0.8.2

# CORS
python-multipart==0.0.20

# Logging and Monitoring
python-json-logger==3.2.1

# Database (PostgreSQL + SQLAlchemy)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
greenlet>=3.0.0

# Redis (Event Bus)
redis[hiredis]>=5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0


